requirements.txt
camelot-py[cv]
pandas
opencv-python
ghostscript
pymupdf